requirements.txt
dj-database-url==0.5.0
Django==2.2.4
django-colorfield==0.1.15
gunicorn==19.9.0
Pillow==6.1.0
psycopg2==2.8.3
pytz==2019.2
sqlparse==0.3.0
whitenoise==4.1.3
